// src/pages/Nutrition/types/nutrition.types.ts
export interface NutritionType {
    id: number;
    PRODUCT_NAME: string;
    PRODUCT_ORIGIN_NAME: string;
    PRODUCT_CODE: string;
    BASE_QY: number;
    KCAL_QY: number;
    PROTEIN_QY: number;
    FAT_QY: number;
    CARBOH_QY: number;
    SUGAR_QY: number;
    DIETARY_FIBER_QY: number;
    CALCIUM_QY: number;
    IRON_QY: number;
    PHOSPHORUS_QY: number;
    POTASSIUM_QY: number;
    NATRIUM_QY: number;
    VITAMIN_A_QY: number;
    B_CAROTENE_QY: number;
    THIAMIN: number;
    VITAMIN_B2_QY: number;
    VITAMIN_B3_QY: number;
    VITAMIN_C_QY: number;
    VITAMIN_D_QY: number;
    CHOLESTEROL_QY: number;
    SATURATED_FAT_QY: number;
    TRANS_FAT_QY: number;
    DATA_ORIGIN: string;
    FOOD_WEIGHT: number;
    COMPANY_NAME: string;
    CREATION_DATE: string;
    BASE_DATE: string;
  }